+++ src/lib/components/Table/Table.d.ts
@@ -34,6 +34,12 @@ export interface TableProps {
   /** Default columns as being sorted */
   defaultSorted?: Array<DefaultSort>;
 
+  /** Hides the pagination / puts all rows on single page */
+  hidePagination?: boolean;
+
+  /** Hides the search box */
+  hideSearch?: boolean;
+
   /** Menu to render */
   menu?: React.ReactNode;
 
@@ -48,6 +54,9 @@ export interface TableProps {
 
   /** Title to render */
   title?: string;
+
+  /** Ref to attach to table */
+  tableRef?: React.Ref;
 }
 
 declare class Table extends React.Component<TableProps, {}> {}