+++ src/lib/components/Table/Table.js
@@ -48,12 +48,14 @@ class Table extends React.Component {
       cellEdit,
       className,
       defaultSorted,
+      hidePagination,
       hideSearch,
       menu,
       rowClasses,
       rowEvents,
       selectRow,
       title,
+      tableRef,
     } = this.props;
 
     const { columns, ...rest } = baseProps;
@@ -74,11 +76,12 @@ class Table extends React.Component {
       defaultSorted,
       hover: true,
       noDataIndication: 'No Entries',
-      pagination,
       rowClasses,
       rowEvents,
+      ...(!hidePagination && { pagination }),
       ...(selectRow && { selectRow }),
       ...(cellEdit && { cellEdit }),
+      ...(tableRef && { ref: tableRef }),
     };
 
     return (
@@ -136,6 +139,9 @@ Table.propTypes = {
     }),
   ),
 
+  /** Hides the pagination / puts all rows on single page */
+  hidePagination: T.bool,
+
   /** Hides the search box */
   hideSearch: T.bool,
 
@@ -155,18 +161,25 @@ Table.propTypes = {
 
   /** Title to render */
   title: T.string,
+
+  /** Ref to attach to table */
+  tableRef: T.shape({
+    current: T.object,
+  }),
 };
 
 Table.defaultProps = {
   cellEdit: null,
   className: null,
   defaultSorted: null,
+  hidePagination: false,
   hideSearch: false,
   menu: null,
   rowClasses: null,
   rowEvents: null,
   selectRow: null,
   title: '',
+  tableRef: null,
 };